requirements.txt
anthropic
assemblyai
xhtml2pdf
matplotlib
click
python-dotenv
